requirements.txt
allure-pytest>=2.13.5
pytest>=8.3.5
pytest-ignore-flaky>=2.2.1
pytest-repeat>=0.9.3
pytest-retry>=1.7.0
pytest-timeout>=2.3.1
selenium>=4.29.0
urllib3>=2.3.0
PyYAML>=6.0.2
dotmap~=1.3.30
allure-python-commons~=2.13.5